+++ src/Forms/FormShortcode.php
@@ -34,8 +34,6 @@ class FormShortcode {
 		$this->forms_module = $forms_module;
 
 		add_shortcode( 'pronamic_payment_form', array( $this, 'shortcode_form' ) );
-
-		add_action( 'init', array( $this, 'shortcode_ui_register' ) );
 	}
 
 	/**
@@ -64,45 +62,5 @@ class FormShortcode {
 		}
 
 		return $this->forms_module->get_form_output_by_id( $atts['id'] );
-	}
-
-	/**
-	 * Shortcode user interface register.
-	 *
-	 * @return void
-	 */
-	public function shortcode_ui_register() {
-		if ( ! function_exists( 'shortcode_ui_register_for_shortcode' ) ) {
-			return;
-		}
-
-		shortcode_ui_register_for_shortcode(
-			'pronamic_payment_form',
-			array(
-				// Display label. String. Required.
-				'label'         => __( 'Payment Form', 'pronamic_ideal' ),
-
-				// Icon/attachment for shortcode. Optional. src or dashicons-$icon. Defaults to carrot.
-				'listItemImage' => 'dashicons-money',
-
-				// Available shortcode attributes and default values. Required. Array.
-				// Attribute model expects 'attr', 'type' and 'label'.
-				// Supported field types: text, checkbox, textarea, radio, select, email, url, number, and date.
-				'attrs'         => array(
-
-					array(
-						'label'    => __( 'Select Payment Form', 'pronamic_ideal' ),
-						'attr'     => 'id',
-						'type'     => 'post_select',
-						'query'    => array(
-							'post_type' => 'pronamic_pay_form',
-						),
-						'multiple' => false,
-					),
-
-				),
-
-			)
-		);
 	}
 }